Factor out BulkUpload from single book uploader
